:lipstick: style : 네비바 스크롤 변경 #109 및 팝업 모바일화 #108 - 네비바 가독성을 위헤 >> 표시를 넣어줬습니다. - 이벤트 폼 및 이벤트 상세정보 창을 모바일화 하였습니다.
