+++ public/stylesheets/components/nav.css
@@ -4,11 +4,11 @@
   display: none;
 }
 
-.navbar-toggle-all.category-item {
+.navbar-toggle-all.box-wrapper {
   display: none;
 }
 
-.navbar-toggle-my.category-item {
+.navbar-toggle-my.box-wrapper {
   display: none;
 }
 
@@ -87,7 +87,7 @@ nav {
   cursor: pointer;
 }
 
-.category-item {
+.box-wrapper {
   align-items: center;
   border-radius: 5px;
   box-shadow: 2px 2px var(--nav_box_shadow);
@@ -102,18 +102,18 @@ nav {
   width: 20%;
 }
 
-.category-item:hover {
+.box-wrapper:hover {
   filter: contrast(1.2);
   transition: filter 0.3s ease-in-out;
 }
 
-.category-item > div {
+.box-wrapper > div {
   font-size: 0.87em;
   padding: 0 0 0 10px;
   width: 70%;
 }
 
-.category-item > i {
+.box-wrapper > i {
   background: var(--white);
   display: block;
   font-size: 1em;
