+++ public/stylesheets/components/popups/event_details_mobile.css
@@ -1,20 +1,22 @@
 .layout-details {
   background: rgba(248, 248, 248, 0.9);
   display: none;
-  height: 100vh;
+  height: 100%;
   left: 0;
   place-items: center;
   position: fixed;
   top: 0;
-  width: 100vw;
+  width: 100%;
+  word-break: break-all;
   z-index: 3;
 }
 
 .details-content {
   background: var(--white);
   display: grid;
-  max-width: 800px;
-  width: 40%;
+  max-height: 98%;
+  width: 60%;
+  max-width: 432px;
 }
 
 .category-button {
@@ -22,38 +24,11 @@
   position: relative;
 }
 
-.duration-location {
-  align-items: center;
-  background: #01a2a4;
+.details-category {
   color: var(--white);
-  display: flex;
-  flex-wrap: wrap;
-  line-height: 18px;
-  min-height: 40px;
-  padding: 20px 0;
-}
-
-.duration-location > div > i {
-  font-size: 20px;
-  vertical-align: -4.5px;
-}
-
-.details-duration {
-  padding: 0 0 0 10%;
-  width: 150px;
-}
-
-.details-location {
-  padding: 10px 10% 0 10%;
-}
-
-.details-exit-button {
   line-height: 30px;
   padding: 5px 0;
-  position: absolute;
-  right: 0px;
-  top: 0px;
-  vertical-align: -1;
+  width: 100%;
 }
 
 .details-category,
@@ -63,11 +38,50 @@
   color: var(--white);
 }
 
-.details-category {
+.details-title {
+  line-height: 1.25em;
+  padding: 1% 2.5em;
+  max-height: 3.75em;
+  overflow: auto;
+  -ms-overflow-style: none;
+  scrollbar-width: none;
+}
+
+.details-title::-webkit-scrollbar {
+  display: none;
+}
+
+.duration-location {
+  align-items: center;
+  background: #01a2a4;
   color: var(--white);
+  display: flex;
+  flex-wrap: wrap;
+  min-height: 40px;
+}
+
+.duration-location > div > i {
+  font-size: 20px;
+  vertical-align: -4.5px;
+}
+
+.details-duration {
+  padding: 3% 0 0 10%;
+  width: 140px;
+}
+
+.details-location {
+  line-height: 20px;
+  padding: 10px 10%;
+}
+
+.details-exit-button {
   line-height: 30px;
   padding: 5px 0;
-  width: 100%;
+  position: absolute;
+  right: 0px;
+  top: 0px;
+  vertical-align: -1;
 }
 
 .details-exit-button {
@@ -83,12 +97,6 @@
   color: var(--details_exit_grey);
 }
 
-.details-title {
-  line-height: 1.25em;
-  padding: 1% 2.5em;
-  word-break: keep-all;
-}
-
 .details-beginat {
   line-height: 35px;
 }
@@ -98,6 +106,7 @@
   display: flex;
   flex-direction: column;
   min-height: 80px;
+  height: 80%;
   padding: 10px 0;
   place-items: center;
   white-space: pre-wrap;
@@ -109,7 +118,7 @@
 .details-topic,
 .details-details {
   line-height: normal;
-  padding: 10px 0px;
+  padding: 5 0 10px 0px;
   width: 80%;
   white-space: pre-wrap;
   word-break: break-word;
@@ -119,12 +128,12 @@
 .details-pic-label,
 .details-topic-label,
 .details-details-label {
-  height: 30px;
-  line-height: 30px;
+  height: 1.75em;
+  line-height: 1.75em;
 }
 
 .details-topic-content {
-  max-height: 100px;
+  max-height: 4em;
   overflow: auto;
   -ms-overflow-style: none; /* for ie and edge */
   scrollbar-width: none; /* for firefox */
@@ -136,8 +145,13 @@
   display: none;
 }
 
+.details-details {
+  display: flex;
+  flex-direction: column;
+}
+
 .details-details-content {
-  max-height: 280px;
+  max-height: 8em;
   overflow: auto;
   -ms-overflow-style: none; /* for ie and edge */
   scrollbar-width: none; /* for firefox */
@@ -152,9 +166,10 @@
   align-items: flex-start;
   display: flex;
   justify-content: space-between;
-  flex-wrap: wrap;
+  justify-self: center;
   height: 50px;
-  padding: 10px 50px 30px;
+  width: 80%;
+  margin-bottom: 20px;
 }
 
 .details-myevent-button {
@@ -162,10 +177,8 @@
   border: none;
   color: var(--white);
   cursor: pointer;
-  float: right;
   height: 50px;
-  line-height: 50px;
-  width: 120px;
+  width: minmax(30%, 120px);
 }
 
 .details-myevent-button:hover {
