+++ public/stylesheets/presets/category.css
@@ -3,7 +3,7 @@
 .month-label-multi.lecture,
 .month-label-end.lecture,
 .lecture .single-category,
-.category-item.lecture,
+.box-wrapper.lecture,
 .category-lecture {
   background: var(--category_lecture);
 }
@@ -11,7 +11,7 @@
 .month-label-multi.exam,
 .month-label-end.exam,
 .exam .single-category,
-.category-item.exam,
+.box-wrapper.exam,
 .category-exam {
   background: var(--category_exam);
 }
@@ -19,7 +19,7 @@
 .month-label-multi.contest,
 .month-label-end.contest,
 .contest .single-category,
-.category-item.contest,
+.box-wrapper.contest,
 .category-contest {
   background: var(--category_contest);
 }
@@ -27,7 +27,7 @@
 .month-label-multi.conference,
 .month-label-end.conference,
 .conference .single-category,
-.category-item.conference,
+.box-wrapper.conference,
 .category-conference {
   background: var(--category_conference);
 }
@@ -35,7 +35,7 @@
 .month-label-multi.community,
 .month-label-end.community,
 .community .single-category,
-.category-item.community,
+.box-wrapper.community,
 .category-community {
   background: var(--category_community);
 }
