+++ views/eventList.ejs
@@ -1,5 +1,5 @@
 <%- contentFor('scriptFiles') %>
-<script defer type="module" src="/javascripts/eventList/generateEventList.js"></script>
+<script defer type="module" src="/javascripts/eventList/setLayoutForEventList.js"></script>
 <script defer type="module" src="/javascripts/eventForm/checkTime.js"></script>
 <script defer type="module" src="/javascripts/eventForm/clickNewEventButton.js"></script>
 <script defer type="module" src="/javascripts/eventForm/eventFormDisplay.js"></script>
