+++ views/index.ejs
@@ -1,6 +1,5 @@
 <%- contentFor('scriptFiles') %>
 <script defer type="module" src="/javascripts/nav/nav.js"></script>
-<script defer type="module" src="/javascripts/nav/navToggle.js"></script>
 <script defer type="module" src="/javascripts/generateCalendars/month/generateMonth.js"></script>
 <script defer type="module" src="/javascripts/generateCalendars/month/changeMonth.js"></script>
 <script defer type="module" src="/javascripts/generateCalendars/month/resizeHandler.js"></script>
